src/modules/cars/useCases/createCars/CreateCarUseCase.ts
import { Cars } from '@modules/cars/infra/typeorm/entities/Cars';
import { ICreateCarDTO } from '@modules/cars/dtos/ICreateCarDTO';
import { ICarRepository } from '@modules/cars/repositories/ICarRepository';
import { AppError } from '@shared/errors/AppError';
import { inject, injectable } from 'tsyringe';

interface IRequest {
  name: string;
  description: string;
  daily_rate: number;
  license_plate: string;
  fine_amount: number;
  brand: string;
  category_id: string;
}

@injectable()
class CreateCarUseCase {
  constructor(
    @inject('CarsRepository')
    private carsRepository: ICarRepository,
  ) {}

  async execute({
    name,
    description,
    daily_rate,
    license_plate,
    fine_amount,
    brand,
    category_id,
  }: IRequest): Promise<Cars> {
    const carAlreadyExists = await this.carsRepository.findByLicensePlate(
      license_plate,
    );
    if (carAlreadyExists) {
      throw new AppError('Car already exists');
    }
    const cars = await this.carsRepository.create({
      name,
      description,
      daily_rate,
      license_plate,
      fine_amount,
      brand,
      category_id,
    });
    return cars;
  }
}

export { CreateCarUseCase };
